Bug fixes and code rearrangement.
Changed a section where 2 dates are compared so that they properly equal
each other in ACF and Railo.

Modified the "dialog" and "view" selectors to be dynamic names i.e.
"dialog-1" so that modals are generated for each event. Also merged the
modal.js code into cf-calendar.cfm so that the selectors could be named
properly as well.

Set class ".window" to display: none; so modal content is still hidden
after above changes.